+++ ZenPacks/zenoss/ZenPackLib/tests/test_datasource_defaults_inheritance.py
@@ -12,32 +12,14 @@
 """
     datasource inheritance custom Datasource with DEFAULTS (ZEN-24083)
 """
-# Zenoss Imports
-import Globals  # noqa
-from Products.ZenUtils.Utils import unused
-unused(Globals)
 
-# stdlib Imports
-from Products.ZenTestCase.BaseTestCase import BaseTestCase
+from ZenPacks.zenoss.ZenPackLib.tests.ZPLTestBase import ZPLTestBase
 
-# zenpacklib Imports
-from ZenPacks.zenoss.ZenPackLib.tests.ZPLTestHarness import ZPLTestHarness
-
-
-def aggregator_installed():
-    """Return True if AggregatingDataSource is installed"""
-    try:
-        from ZenPacks.zenoss.CalculatedPerformance.datasources.AggregatingDataSource import AggregatingDataSource
-    except ImportError:
-        pass
-    else:
-        return True
-    return False
 
 YAML_DOC = """
-name: ZenPacks.zenoss.ZenPackLib
+name: ZenPacks.zenoss.CalculatedPerformance
 device_classes:
-  /Device:
+  /Devices:
     templates:
       TESTAGGFAIL:
         description: you should see that targetDatsource does not set
@@ -62,22 +44,23 @@ device_classes:
 """
 
 
-class TestZen24083(BaseTestCase):
+class TestDatasourceInheritance(ZPLTestBase):
     """Test fix for ZEN-24083
 
        datasource inheritance custom Datasource with DEFAULTS
     """
+    yaml_doc = YAML_DOC
+    use_dmd = True
+    disableLogging = False
 
     def test_datasource_inheritance(self):
-        if aggregator_installed():
-            z = ZPLTestHarness(YAML_DOC)
-            z.connect()
-            self.assertTrue(z.check_templates_vs_yaml(), "Template objects do not match YAML")
-            self.assertTrue(z.check_templates_vs_specs(), "Template objects do not match Spec")
+        if self.z.zenpack_installed():
+            self.assertTrue(self.z.check_templates_vs_yaml(), "Template objects do not match YAML")
+            self.assertTrue(self.z.check_templates_vs_specs(), "Template objects do not match Spec")
             # check properties on dummy template
-            dcs = z.cfg.device_classes.get('/Device')
+            dcs = self.z.cfg.device_classes.get('/Devices')
             tcs = dcs.templates.get('TESTAGGFAIL')
-            t = tcs.create(z.dmd, False)
+            t = tcs.create(self.z.dmd, False)
             for ds in t.datasources():
                 # standard paramenter
                 self.assertEquals(ds.component, '${here/id}', 'datasource property (component) was not inherited from DEFAULTS')
@@ -86,14 +69,15 @@ class TestZen24083(BaseTestCase):
                 # test local override of DEFAULTS
                 if ds.id == 'agg_in_octets':
                     self.assertEquals(ds.targetDataPoint, 'aggifHCOutOctets', 'datasource property (targetDataPoint) DEFAULTS override was not set')
-
+        else:
+            self.log.warn('Skipping test_integer_threshold since ZenPacks.zenoss.CalculatedPerformance not installed.')
 
 
 def test_suite():
     """Return test suite for this module."""
     from unittest import TestSuite, makeSuite
     suite = TestSuite()
-    suite.addTest(makeSuite(TestZen24083))
+    suite.addTest(makeSuite(TestDatasourceInheritance))
     return suite
 
 if __name__ == "__main__":